feat(#150): add more javadocs
