+++ src/main/java/org/eolang/opeo/decompilation/InstructionHandler.java
@@ -34,5 +34,5 @@ public interface InstructionHandler {
      * Handle instruction.
      * @param state Current instruction to handle together with operand stack and variables.
      */
-    void handle(final MachineState state);
+    void handle(final DecompilerState state);
 }
