+++ src/main/java/org/eolang/opeo/decompilation/handlers/AddHandler.java
@@ -28,7 +28,7 @@ import org.eolang.opeo.ast.AstNode;
 import org.eolang.opeo.ast.Attributes;
 import org.eolang.opeo.ast.Opcode;
 import org.eolang.opeo.decompilation.InstructionHandler;
-import org.eolang.opeo.decompilation.MachineState;
+import org.eolang.opeo.decompilation.DecompilerState;
 import org.objectweb.asm.Opcodes;
 
 /**
@@ -44,7 +44,7 @@ public class AddHandler implements InstructionHandler {
     }
 
     @Override
-    public void handle(final MachineState state) {
+    public void handle(final DecompilerState state) {
         if (state.instruction().opcode() == Opcodes.IADD) {
             final AstNode right = state.stack().pop();
             final AstNode left = state.stack().pop();
