+++ src/main/java/org/eolang/opeo/decompilation/handlers/PutFieldHnadler.java
@@ -28,7 +28,7 @@ import org.eolang.opeo.ast.Attributes;
 import org.eolang.opeo.ast.Field;
 import org.eolang.opeo.ast.FieldAssignment;
 import org.eolang.opeo.decompilation.InstructionHandler;
-import org.eolang.opeo.decompilation.MachineState;
+import org.eolang.opeo.decompilation.DecompilerState;
 
 /**
  * Putfield instruction handler.
@@ -38,7 +38,7 @@ import org.eolang.opeo.decompilation.MachineState;
 public class PutFieldHnadler implements InstructionHandler {
 
     @Override
-    public void handle(final MachineState state) {
+    public void handle(final DecompilerState state) {
         final AstNode value = state.stack().pop();
         final String name = (String) state.operand(1);
         final String owner = (String) state.operand(0);
