+++ src/main/java/org/eolang/opeo/decompilation/handlers/ReturnHandler.java
@@ -25,7 +25,7 @@ package org.eolang.opeo.decompilation.handlers;
 
 import org.eolang.opeo.ast.Opcode;
 import org.eolang.opeo.decompilation.InstructionHandler;
-import org.eolang.opeo.decompilation.MachineState;
+import org.eolang.opeo.decompilation.DecompilerState;
 
 /**
  * Return instruction handler.
@@ -40,7 +40,7 @@ public class ReturnHandler implements InstructionHandler {
     }
 
     @Override
-    public void handle(final MachineState state) {
+    public void handle(final DecompilerState state) {
         state.stack().push(
             new Opcode(
                 state.instruction().opcode(),
