+++ src/main/java/org/eolang/opeo/decompilation/handlers/SubstractionHandler.java
@@ -28,7 +28,7 @@ import org.eolang.opeo.ast.Attributes;
 import org.eolang.opeo.ast.Opcode;
 import org.eolang.opeo.ast.Substraction;
 import org.eolang.opeo.decompilation.InstructionHandler;
-import org.eolang.opeo.decompilation.MachineState;
+import org.eolang.opeo.decompilation.DecompilerState;
 import org.objectweb.asm.Opcodes;
 
 /**
@@ -44,7 +44,7 @@ public class SubstractionHandler implements InstructionHandler {
     }
 
     @Override
-    public void handle(final MachineState state) {
+    public void handle(final DecompilerState state) {
         if (state.instruction().opcode() == Opcodes.ISUB) {
             final AstNode right = state.stack().pop();
             final AstNode left = state.stack().pop();
